Revert splitting out code from handlers
